Security/NonceVerification: disregard nonce check in PHP 7.4 arrow function when searching
As an arrow function is a callback and we don't trace whether a callback is called, any nonce check found within an arrow function should be disregarded.
This mirrors the sniff behaviour with regards to closures, meaning nonce checks in nested arrow functions and in nested closures will now be treated the same.

Includes test.