Security/EscapeOutput: fix up after refactor
The refactor in the previous few commits "broke" the preliminary check for a non-parenthesized ternary ìn a `print` statement.

There were no tests covering this.

I'm fixing this now by moving that particular code block out of the `switch` and executing it for both `print`, `echo` and `<?=` (though not for `exit`/`die` statements as those will always have parentheses, otherwise there will be nothing to check).

I'm also extracting the actual logic for finding the ternary to a separate method, as this logic will be expanded in follow-up commits.

Includes adding tests safeguarding the logic.