Security/EscapeOutput: bug fix - allow for nested statements with ternaries
As things were, the check for statements with ternaries without wrapping parentheses, did not take into account that the whole statement could be nested within parentheses if the statement is a `print`.

This commit updates the code block handling the skipping of the condition for a ternary without wrapping parentheses to allow for the statement being _nested in_ parentheses.

Includes adding tests safeguarding the fix.