Security/EscapeOutput: bug fix - allow non-matching parentheses
As things were, the prelim check for statements with ternaries without wrapping parentheses, did not take into account that there could be a parentheses at the start and at the end of the statement, which didn't belong together.
In that case, the preliminary ternary check should still be executed.

Fixed now.

Includes adding tests safeguarding the fix.

Fixes 677

This also fixes the third case reported in 1507