Security/EscapeOutput: add support for PHP 8.0+ match expressions [2]
PHP 8.0 introduced match expressions, which can be used inline, which means they can be used in an output generating statement.

This second commit adds logic to correctly check _inside_ a match expression.

In a `match`, there can be multiple matching conditions before the double arrow and one statement to be returned after the double arrow.
So basically, we should ignore everything _before_ the double arrow as it is effectively a comparison.
And we should verify that everything after the double arrow - up to the first comma at the same "level" as the arrow - is escaped correctly.

Now as the conditions, when there is more than one, are also comma-separated, the parsing of a `match` expression is non-trivial.

With this in mind, I'm introducing a helper method to do the parsing. This helper method will pass each individually identified "return value" part to the `check_code_is_escaped()` method for stand-alone analysis.

Once the method is finished analyzing the `match` expression, it will hand back off to the original call to `check_code_is_escaped()` and continue analyzing the code after the `match` expression.

Note: I'm considering adding a utility method to PHPCSUtils for the parsing of `match` expressions. If and when that method is available, we should be able to simplify the code I've now added to the sniff.

Includes tests.