Security/EscapeOutput: bug fix - expand the "safe components" list
Operator tokens in PHP are not insecure, it is the **_operands_** which are potentially insecure.

As things were, we've seen all sorts of messages in bugs reported, where the sniff would report on `&&` or `===` tokens, which is nonsensical.

The "safe components" list already contained a number of operators, most notably the arithmetic operators (though the PHP 5.6+ `**` (`T_POW`) operator was missing).

I'm proposing to expand this list with a much larger list of operators using the predefined token arrays from PHPCS itself + one of the PHPCSUtils token arrays + the `T_BOOLEAN_NOT` (`!`) operator.

With this expansion, we shouldn't be seeing anymore issues reported for the sniff flagging operators, which should make debugging the issues reported more straight forward.

The only operators I have not (yet) included in the list are the assignment operators as assignments within an output statement should be exceedingly rare anyway.

Note: in all honesty, the sniff should take operator precedence into account properly and, for instance, disregard a `$var` or function call after a boolean operator as that is automatically cast to a boolean. Similarly, the operands before and after an arithmetic operator will always be cast to numeric. Having said that, that's a rabbit hole for a future release as there's enough to fix in the sniff without going down that path.

Includes unit tests, all of which would throw an error on the operator prior to this change.