[SanitizingFunctions|EscapingFunctions]Trait: add wp_kses_one_attr() to function lists
This function was introduced in WP 4.2.3.

I've done a cursory look-over of the code, but I have not gone down the rabbit-hole of examining all functions involved (and yes, `wp_kses_one_attr()` calls a _lot_ of functions under the hood).

Having said that, anastis did do a more detailed analysis and included their write-up of it in the ticket, so I'm satisfied that this function can be added to the list.

Based on this write-up, I believe the function should be added to both the `$sanitizingFunctions` list as well as the `$escapingFunctions` list.

Refs:
* https://developer.wordpress.org/reference/functions/wp_kses_one_attr/
* WordPress/WordPress-Coding-Standards 1765#issuecomment-512192213

Fixes 1765