Security/EscapeOutput: improve handling of heredocs, include PHP 7.3+ flexible heredocs
As things were, when the sniff would encounter a heredoc, it would throw an error on the heredoc opener (`<<<EOD`) token and be done with it.

However, heredocs _can_, but don't necessarily always contain interpolated variables and expressions and when the heredoc doesn't contain any interpolation, it should be treated the same as any other plain text string.

This commit fixes the sniff to handle heredocs properly and throw errors only on heredoc lines containing interpolation.

Includes tests.