SanitizationHelperTrait::is_sanitized(): add some defensive coding
As this method is now stand-alone, we'd better make sure the token being examined exists.

Note: no need to do the same for the `is_only_sanitized()` method as the first thing that method calls is this method and if this method returns `false`, the `is_only_sanitized()` method will too.